fix: bacnet_action_command_decode() truncates Post_Delay to 8 bits (#1423)
BACNET_ACTION_LIST.Post_Delay is declared uint32_t (bacaction.h) and
bacnet_action_command_encode() writes the full 32-bit value for the
postDelay [6] context tag. However bacnet_action_command_decode()
stored the decoded value through a stray (uint8_t) cast, so any
postDelay >= 256 was silently truncated on decode (e.g. 1000 -> 232,
65536 -> 0), corrupting round-tripped Command object Action list
entries.

Change the cast to (uint32_t) to match the field's declared type and
the encoder's behavior.

+++ src/bacnet/bacaction.c
@@ -487,7 +487,7 @@ int bacnet_action_command_decode(
     if (len > 0) {
         apdu_len += len;
         if (entry) {
-            entry->Post_Delay = (uint8_t)unsigned_value;
+            entry->Post_Delay = (uint32_t)unsigned_value;
         }
     } else {
         /* wrong tag or malformed - optional - skip apdu_len increment */
